pyproject.toml
[tool.poetry]
name = "PolyMaG_gendiff"
version = "0.6.2"
description = ""
authors = ["Maxim P"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest-cov = "^2.8.1"
pytest = "^5.3.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
